+++ Sources/nostr/Event/Tags/Event + Tags.swift
@@ -1,0 +1,3 @@
+extension Event {
+	public typealias Tags = [Tag]
+}