+++ Sources/nostr/Protocols/DYTAG/DYTAG_encodable.swift
@@ -1,3 +1,0 @@
-public protocol DYTAG_encodable {
-	var dytag_nameVals:[Character:[String]] { get }
-}