+++ Sources/nostr/Protocols/TAG/TAG_encodable.swift
@@ -1,3 +1,0 @@
-public protocol TAG_encodable {
-	var tag_nameVals:[String:[String]] { get }
-}